requirements.txt
streamlit==1.40.1
PyPDF2==3.0.1
python-docx==1.1.2
transformers==4.46.3
sentence-transformers==3.3.1
pinecone-client==5.0.1
groq==0.11.0
torch==2.5.1